dom/tests/mochitest/geolocation/test_timerRestartWatch.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=526326
-->
<head>
  <title>Test for watchPosition </title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=526326">Mozilla Bug 526326</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var watchID;
var hasAccepted = false;

function errorCallback(err) {
  ok(err.code == err.TIMEOUT, "ensure error is a timeout.");
  resume_geolocationProvider();
  SimpleTest.finish();
}

function successCallback(position) {
  ok(hasAccepted, "Ensure that accept was pressed");

  // Now that we got a success callback, lets try to ensure
  // that we get a timeout error.

  stop_geolocationProvider();
}

function accept() {
  hasAccepted = true;
  clickNotificationButton(kAcceptButton);
}

/** Test for Bug  **/

SimpleTest.waitForExplicitFinish();


var options = {
    maximumAge: 0,
    timeout: 1000
};

watchID = navigator.geolocation.watchPosition(successCallback, errorCallback, options);

setTimeout(accept, 50);

</script>
</pre>
</body>
</html>
